+++ src/django/planit_data/management/commands/update_fixtures.py
@@ -7,7 +7,9 @@ from django.conf import settings
 FIXTURES = {
     'planit_data': [
         # Organized (filename, model_list)
+        ('climateassessmentregions.json', ['planit_data.climateassessmentregion']),
         ('communitysystems.json', ['planit_data.communitysystem']),
+        ('counties.json', ['planit_data.county']),
         ('concerns.json', ['planit_data.concern']),
         ('defaultrisks.json', ['planit_data.defaultrisk']),
         ('georegions.json', ['planit_data.georegion']),